+++ features/step_definitions/checks.rb
@@ -178,7 +178,7 @@ Then("the container is( still) there/present") do
 end
 
 
-And(/^file ([^\s]+) (?:still|now )?contains$/) do |file_name, expected_content|
+And(/^file ([^\s]+) (?:still |now )?contains$/) do |file_name, expected_content|
   expect(File).to exist(file_name)
   actual_content = File.read(file_name)
   expect(actual_content.strip).to eq(expected_content.strip)