+++ CHANGELOG.md
@@ -13,6 +13,10 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
  - Kill rogue containers with `--kill` (issue #11)
  - Kill containers if setup fails (issue #7)
 
+### Fixes
+
+ - Avoid redundant `docker start`
+
 
 ## [1.1.0] - 2020-06-29